rfactor: extract CacheQueries trait from cache.rs, implement for ClickHouse
refactor: move CacheQueries trait and ClickHouse impl into db module

refactor: move cache e2e tests to db/cache_queries.rs following make_clickhouse_only_test pattern

refactor: split cache tests - CacheQueries tests in db/cache_queries.rs, full e2e tests in cache.rs
