+++ tensorzero-core/src/cache.rs
@@ -1,9 +1,8 @@
-use std::collections::HashMap;
 use std::future::Future;
 use std::sync::Arc;
 
 use crate::config::OtlpConfig;
-use crate::db::clickhouse::{ClickHouseConnectionInfo, TableName};
+use crate::db::cache::CacheQueries;
 use crate::embeddings::{Embedding, EmbeddingModelResponse, EmbeddingRequest};
 use crate::error::{Error, ErrorDetails, warn_discarded_cache_write};
 use crate::inference::types::{
@@ -212,18 +211,7 @@ impl ModelProviderRequest<'_> {
     }
 }
 
-// The full row written to ClickHouse
-#[derive(Debug, Serialize)]
-struct FullCacheRow<T: CacheOutput> {
-    short_cache_key: u64,
-    long_cache_key: String,
-    // We flatten this so that the fields map directly to columns in the ClickHouse table
-    #[serde(flatten)]
-    data: CacheData<T>,
-}
-
-/// The underlying cached input/output data. These are the fields that we actually retrieve from
-/// ClickHouse when going a cache fetch
+/// The underlying cached input/output data.
 #[derive(Debug, Deserialize, Serialize)]
 pub struct CacheData<T: CacheOutput> {
     pub output: T,
@@ -317,26 +305,33 @@ pub struct StreamingCacheData {
     pub chunks: Vec<CachedProviderInferenceResponseChunk>,
 }
 
-fn spawn_maybe_cache_write<T: Serialize + CacheOutput + Send + Sync + 'static>(
-    row: FullCacheRow<T>,
-    clickhouse_client: ClickHouseConnectionInfo,
+fn spawn_maybe_cache_write<
+    T: Serialize + CacheOutput + Send + Sync + 'static,
+    C: CacheQueries + Clone + 'static,
+>(
+    cache_key: CacheKey,
+    cache_data: CacheData<T>,
+    cache_backend: C,
     cache_validation_info: CacheValidationInfo,
 ) {
     spawn_ignoring_shutdown(async move {
-        if row
-            .data
+        if cache_data
             .output
             .should_write_to_cache(cache_validation_info)
             .await
         {
-            if let Err(e) = clickhouse_client
-                .write_batched(&[row], TableName::ModelInferenceCache)
-                .await
-            {
+            let data_json = match serde_json::to_string(&cache_data) {
+                Ok(json) => json,
+                Err(e) => {
+                    tracing::warn!("Failed to serialize cache data: {e}");
+                    return;
+                }
+            };
+            if let Err(e) = cache_backend.cache_write(&cache_key, &data_json).await {
                 tracing::warn!("Failed to write to cache: {e}");
             }
         } else {
-            warn_discarded_cache_write(&row.data.raw_response);
+            warn_discarded_cache_write(&cache_data.raw_response);
         }
     });
 }
@@ -356,24 +351,17 @@ pub struct CacheValidationInfo {
 }
 
 // This doesn't block
-pub fn start_cache_write<T: Serialize + CacheOutput + Send + Sync + 'static>(
-    clickhouse_client: &ClickHouseConnectionInfo,
+pub fn start_cache_write<
+    T: Serialize + CacheOutput + Send + Sync + 'static,
+    C: CacheQueries + Clone + 'static,
+>(
+    cache_backend: &C,
     cache_key: CacheKey,
     cache_data: CacheData<T>,
     cache_validation_info: CacheValidationInfo,
 ) -> Result<(), Error> {
-    let short_cache_key = cache_key.get_short_key()?;
-    let long_cache_key = cache_key.get_long_key();
-    let clickhouse_client = clickhouse_client.clone();
-    spawn_maybe_cache_write(
-        FullCacheRow {
-            short_cache_key,
-            long_cache_key,
-            data: cache_data,
-        },
-        clickhouse_client,
-        cache_validation_info,
-    );
+    let cache_backend = cache_backend.clone();
+    spawn_maybe_cache_write(cache_key, cache_data, cache_backend, cache_validation_info);
     Ok(())
 }
 
@@ -388,16 +376,14 @@ pub struct CachedProviderInferenceResponseChunk {
 }
 
 // This starts a trailing write to the cache (without blocking the http response)
-pub fn start_cache_write_streaming(
-    clickhouse_client: &ClickHouseConnectionInfo,
+pub fn start_cache_write_streaming<C: CacheQueries + Clone + 'static>(
+    cache_backend: &C,
     cache_key: CacheKey,
     chunks: Vec<ProviderInferenceResponseChunk>,
     raw_request: &str,
     usage: &Usage,
     tool_config: Option<ToolCallConfig>,
 ) -> Result<(), Error> {
-    let short_cache_key = cache_key.get_short_key()?;
-    let long_cache_key = cache_key.get_long_key();
     let input_tokens = usage.input_tokens;
     let output_tokens = usage.output_tokens;
     let mut finish_reason = None;
@@ -418,33 +404,30 @@ pub fn start_cache_write_streaming(
             .collect(),
     };
     let raw_request = raw_request.to_string();
-    let clickhouse_client = clickhouse_client.clone();
+    let cache_backend = cache_backend.clone();
     spawn_maybe_cache_write(
-        FullCacheRow {
-            short_cache_key,
-            long_cache_key,
-            data: CacheData {
-                output,
-                raw_request,
-                raw_response: String::new(),
-                input_tokens,
-                output_tokens,
-                finish_reason,
-            },
+        cache_key,
+        CacheData {
+            output,
+            raw_request,
+            raw_response: String::new(),
+            input_tokens,
+            output_tokens,
+            finish_reason,
         },
-        clickhouse_client,
+        cache_backend,
         CacheValidationInfo { tool_config },
     );
     Ok(())
 }
 
 pub async fn embedding_cache_lookup(
-    clickhouse_connection_info: &ClickHouseConnectionInfo,
+    cache_backend: &impl CacheQueries,
     request: &EmbeddingModelProviderRequest<'_>,
     max_age_s: Option<u32>,
 ) -> Result<Option<EmbeddingModelResponse>, Error> {
     let result = cache_lookup_inner::<EmbeddingCacheData>(
-        clickhouse_connection_info,
+        cache_backend,
         request.get_cache_key()?,
         max_age_s,
     )
@@ -453,12 +436,12 @@ pub async fn embedding_cache_lookup(
 }
 
 pub async fn cache_lookup(
-    clickhouse_connection_info: &ClickHouseConnectionInfo,
+    cache_backend: &impl CacheQueries,
     request: ModelProviderRequest<'_>,
     max_age_s: Option<u32>,
 ) -> Result<Option<ModelInferenceResponse>, Error> {
     let result = cache_lookup_inner::<NonStreamingCacheData>(
-        clickhouse_connection_info,
+        cache_backend,
         request.get_cache_key()?,
         max_age_s,
     )
@@ -469,16 +452,11 @@ pub async fn cache_lookup(
 }
 
 pub async fn cache_lookup_streaming(
-    clickhouse_connection_info: &ClickHouseConnectionInfo,
+    cache_backend: &impl CacheQueries,
     request: ModelProviderRequest<'_>,
     max_age_s: Option<u32>,
 ) -> Result<Option<StreamResponse>, Error> {
-    let result = cache_lookup_inner(
-        clickhouse_connection_info,
-        request.get_cache_key()?,
-        max_age_s,
-    )
-    .await?;
+    let result = cache_lookup_inner(cache_backend, request.get_cache_key()?, max_age_s).await?;
     Ok(result.map(|result| {
         StreamResponse::from_cache(
             result,
@@ -489,70 +467,22 @@ pub async fn cache_lookup_streaming(
 }
 
 pub async fn cache_lookup_inner<T: CacheOutput + DeserializeOwned>(
-    clickhouse_connection_info: &ClickHouseConnectionInfo,
+    cache_backend: &impl CacheQueries,
     cache_key: CacheKey,
     max_age_s: Option<u32>,
 ) -> Result<Option<CacheData<T>>, Error> {
-    // NOTE: the short cache key is just so the ClickHouse index can be as efficient as possible
-    // but we always check against the long cache key before returning a result
-    let short_cache_key = cache_key.get_short_key()?.to_string();
-    let long_cache_key = cache_key.get_long_key();
-    // The clickhouse query args look like rust format string args, but they're not.
-    let query = if max_age_s.is_some() {
-        r"
-            SELECT
-                output,
-                raw_request,
-                raw_response,
-                input_tokens,
-                output_tokens,
-                finish_reason
-            FROM ModelInferenceCache
-            WHERE short_cache_key = {short_cache_key:UInt64}
-                AND long_cache_key = {long_cache_key:String}
-                AND timestamp > subtractSeconds(now(), {lookback_s:UInt32})
-            ORDER BY timestamp DESC
-            LIMIT 1
-            FORMAT JSONEachRow
-        "
-    } else {
-        r"
-            SELECT
-                output,
-                raw_request,
-                raw_response,
-                input_tokens,
-                output_tokens,
-                finish_reason
-            FROM ModelInferenceCache
-            WHERE short_cache_key = {short_cache_key:UInt64}
-                AND long_cache_key = {long_cache_key:String}
-            ORDER BY timestamp DESC
-            LIMIT 1
-            FORMAT JSONEachRow
-        "
-    };
-    let mut query_params = HashMap::from([
-        ("short_cache_key", short_cache_key.as_str()),
-        ("long_cache_key", long_cache_key.as_str()),
-    ]);
-    let lookback_str;
-    if let Some(lookback) = max_age_s {
-        lookback_str = lookback.to_string();
-        query_params.insert("lookback_s", lookback_str.as_str());
+    let response = cache_backend.cache_lookup(&cache_key, max_age_s).await?;
+    match response {
+        None => Ok(None),
+        Some(json) => {
+            let result: CacheData<T> = serde_json::from_str(&json).map_err(|e| {
+                Error::new(ErrorDetails::Cache {
+                    message: format!("Failed to deserialize output: {e}"),
+                })
+            })?;
+            Ok(Some(result))
+        }
     }
-    let result = clickhouse_connection_info
-        .run_query_synchronous(query.to_string(), &query_params)
-        .await?;
-    if result.response.is_empty() {
-        return Ok(None);
-    }
-    let result: CacheData<T> = serde_json::from_str(&result.response).map_err(|e| {
-        Error::new(ErrorDetails::Cache {
-            message: format!("Failed to deserialize output: {e}"),
-        })
-    })?;
-    Ok(Some(result))
 }
 
 #[cfg(test)]
