+++ tensorzero-core/src/db/mod.rs
@@ -14,6 +14,7 @@ use crate::error::Error;
 use crate::serde_util::{deserialize_option_u64, deserialize_u64};
 
 pub mod batch_inference;
+pub mod cache;
 pub mod clickhouse;
 pub mod datasets;
 pub mod delegating_connection;
